+++ src/OpenTelemetry.Instrumentation.SqlClient/CHANGELOG.md
@@ -6,7 +6,7 @@
   `netstandard2.0` targets. `SetDbStatement` has been removed. Use
   `SetDbStatementForText` to capture command text and stored procedure names on
   .NET Framework.
-  ([#XXXX](https://github.com/open-telemetry/opentelemetry-dotnet/pull/XXXX))
+  ([#3900](https://github.com/open-telemetry/opentelemetry-dotnet/pull/3900))
 
 ## 1.0.0-rc9.9